# openapi.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "NodeJS-FAAS",
      "description": "This project aims to achieve something similar to Function as a Service (FaaS) without the overhead of Docker or Kubernetes, with extremely low infrastructure requirements. \n Github repository link : [Click Here](https://github.com/shreyasnayak/nodejs-faas)",
      "version": "1.0.0"
    },
    "components": {
      "securitySchemes": {
        "BearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    },
    "security": [
      {
        "BearerAuth": []
      }
    ],  
    "paths": {
      "/server/health": {
        "get": {
          "tags": [
            "Server"
          ],
          "summary": "Server health check",
          "responses": {
            "200": {
              "description": "NodeFAAS - Server is running",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "string",
                    "example": "OK"
                  }
                }
              }
            }
          }
        },
        "head": {
          "tags": [
            "Server"
          ],
          "summary": "Server health check",
          "responses": {
            "200": {
              "description": "NodeFAAS - Server is running"
            }
          }
        }
      },
      "/function/list": {
        "get": {
          "tags": [
            "Function"
          ],
          "summary": "Get list of functions",
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      },
                      "code": {
                        "type": "integer"
                      },
                      "message": {
                        "type": "string"
                      },
                      "payload": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "example": {
                      "status": "OK",
                      "code": 200,
                      "message": "",
                      "payload": {
                        "function_ids": ["VmgtJbe8I2", "nmgtkbe8I2"]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/function/{function_id}": {
        "delete": {
          "tags": [
            "Function"
          ],
          "summary": "Delete a function",
          "parameters": [
            {
              "in": "path",
              "name": "function_id",
              "required": true,
              "description": "The alphanumeric ID of the function file to delete",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      },
                      "code": {
                        "type": "integer"
                      },
                      "message": {
                        "type": "string"
                      },
                      "payload": {}
                    },
                    "example": {
                      "status": "OK",
                      "code": 200,
                      "message": "File deleted successfully",
                      "payload": {}
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/function/create": {
        "post": {
          "tags": [
            "Function"
          ],
          "summary": "Upload a function file",
          "requestBody": {
            "required": true,
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "file": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      },
                      "code": {
                        "type": "integer"
                      },
                      "message": {
                        "type": "string"
                      },
                      "payload": {
                        "type": "object"
                      }
                    },
                    "example": {
                      "status": "OK",
                      "code": 200,
                      "message": "File uploaded successfully",
                      "payload": {
                        "function_id": "abcd1234"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/function/execute/{function_id}": {
        "post": {
          "tags": [
            "Function"
          ],
          "summary": "Call a function",
          "parameters": [
            {
              "in": "path",
              "name": "function_id",
              "required": true,
              "description": "The file ID of the function",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "event_name": {
                      "type": "string",
                      "description": "Name of the event"
                    },
                    "event_data": {
                      "type": "object",
                      "description": "Data associated with the event"
                    }
                  },
                  "example": {
                    "event_name": "get_nasa_image",
                    "event_data": {
                      "key": "DEMO_KEY"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      },
                      "code": {
                        "type": "integer"
                      },
                      "message": {
                        "type": "string"
                      },
                      "payload": {
                        "type": "object"
                      }
                    },
                    "example": {
                      "status": "OK",
                      "code": 200,
                      "message": "",
                      "payload": {}
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
  